Remove softfailure if yast2 snapper times out
We have observed sporadic issue that sometimes even yast2 snapper window
is already closed, return code is not yet provided and some activity is
still there. As a part of investigation of bsc#1032831 we have
identified, that yast2 ui tests tend to fail with timeouts and in
observed cases there was some btrfs activity. Traces of if we can see in
serial output. Same behavior is easy to reproduce if one triggers btrfs
balance and tries to start and close snapper, it takes significantly
more time than in normal scenario. Considering the fact that SUT is
usually running on hardware which satisfies minimal requirements,
timeouts are plausible. With this PR we remove workarounds and introduce
increased timeout value to have stable results. Activities regarding
possible btrfs performance regression will be performed as individual
task.