Restore backward-compatibility with CaaSP 1.0
There was a number of changes in the `stack_controller`
which caused backward-compatibility issues in the context of
`CaaSP 1.0 DVD Deployment`. This PR reverts those
changes while it keeps the CaaSP 2.0 compatibility intact.

rebased commits:

* ba213b46f7bac5a017adca46caa3e0bf4bcf69b2
* 7e4a90bfbf0b2a622f2fed1159a634e03c3a70e1

It also uses `caasp` instead of `casp` as the proper naming
scheme. The *QAM* product for CaaSP in `osd` will be named
accordingly.