Quieten original load, extend timeout for JeOS
Wait until load after boot calms down to ~zero and then force-run cron
jobs. For JeOS the load is at this moment around 0.60 and needs 2-3
minutes to calm down.

Even if the "primordial" load is removed from the equasion, the load
after forced-run cron jobs is again considerable enough not to fall fast
enough to make it within 30 second soft timeout - extending it to 3
minutes as well.

Verification runs:
* JeOS: http://assam.suse.cz/tests/579
* SLES: http://assam.suse.cz/tests/581