LTP: Add LTP_EXTRA_CONF_FLAGS for 32-bit compatability mode
This allows --CCFLAGS=-m32 --LDFLAGS=-m32 to be passed to the configure script
causing the LTP to be compiled to a 32-bit binary on x86_64 (the flag is
different or missing on some other architectures). Obviously it allows any
other option to be passed to the configure script as well.

Also tries to install the 32-bit development packages, many of which are
currently missing. To complicate matters; the configure script will mark a
feature as present if only the 64-bit development package is available. So
compilation will fail with cryptic messages about missing libraries or
functions instead of skipping tests which require the missing 32-bit
libraries.

It appears that in most cases the 32-bit version of a library is installed as
part of the normal development package, but is missing a symlink like
library.so -> library.so.1. So we can work around the missing packages if
absolutely necessary.