Check target match with x11_start_program by default in all instances
As checking for a screen with "assert_screen" after "x11_start_program" is the
standard action and it allows to save time because there is a target to look
for this is now the default action saving quite some time for all X11 tests.

Exceptions can still be applied, for example by specifying "valid => 0" on the
call of "x11_start_program".

Validation runs:
* SLE 12 SP1 qam-gnome: http://lord.arch/tests/7609 glxgears failed
  because of differing no_wait, update after newest test
* openSUSE TW yast2_gui: http://lord.arch/tests/7608